telegram_bot/requirements.txt
anyio==4.12.0
asyncio==4.0.0
build==1.3.0
certifi==2025.11.12
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.1
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==25.0
pillow==11.3.0
pyparsing==3.2.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-telegram-bot==22.5
six==1.17.0
TA-Lib==0.6.8
tomli==2.3.0
typing_extensions==4.15.0
websockets==15.0.1
zipp==3.23.0
